sdm845-common: reduce LED blinking frequency under low battery
This is the second attempt of fixing the LED. The previous commit
1e2983474169623e9ad0b7e733921086d6f52c3f breaks notification LED
during charging sometimes. As I said in the previous fix, the phone
does not take pause_lo value properly. I found that the LED blink
frequency varies from 160 bpm to 53 bpm, depending on the value of
pause_lo. It reaches the slowest blink frequency when pause_lo is
2^n-3 where n >= 9. Therefore, 509 is used for the value of
config_notificationsBatteryLedOff, and even 510 will make the LED
go crazy under low battery.

I also boosted the scaled duty_pcts so that single color notifica-
tion LED will not fall under the minimum brightness allowed, as a
compensation for lowering the overall LED brightness.

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Change-Id: If0f7a65aa642bc691b95755a0c9ee6ab9f41eff3